convert shine_enwindow to integers

+++ src/lib/tables.h
@@ -1,8 +1,8 @@
 #ifndef TABLES_H
 #define TABLES_H
 
-extern int    shine_scale_fact_band_index[9][23];
-extern double shine_enwindow[];
+extern int  shine_scale_fact_band_index[9][23];
+extern long shine_enwindow[];
 
 #endif
